Refactor: Move isUrlDirectlyReachable to ServerUrlMapper
- Moves `isUrlDirectlyReachable` from `ResourceViewerViewModel` to `ServerUrlMapper`.
- Injects `DispatcherProvider` into `ServerUrlMapper` to handle coroutine context for network calls.
- Updates `ServerUrlMapperTest` to provide a mock `DispatcherProvider`.
- Removes networking dependencies (`HttpURLConnection`, `URL`) and coroutine `withContext` from the UI layer (`ResourceViewerViewModel`).
