+++ app/src/main/java/org/ole/planet/myplanet/ui/viewer/ResourceViewerViewModel.kt
@@ -2,10 +2,7 @@ package org.ole.planet.myplanet.ui.viewer
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import java.net.HttpURLConnection
-import java.net.URL
 import javax.inject.Inject
-import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.data.auth.AuthSessionUpdater
 import org.ole.planet.myplanet.model.MyLibrary
 import org.ole.planet.myplanet.repository.ResourcesRepository
@@ -27,25 +24,8 @@ class ResourceViewerViewModel @Inject constructor(
         val mapping = serverUrlMapper.processUrl(serverUrl)
         if (mapping.alternativeUrl != null) {
             serverUrlMapper.updateServerIfNecessary(mapping, sharedPrefManager.rawPreferences) { url ->
-                isUrlDirectlyReachable(url)
+                serverUrlMapper.isUrlDirectlyReachable(url)
             }
-        }
-    }
-
-    private suspend fun isUrlDirectlyReachable(url: String): Boolean {
-        return try {
-            withContext(dispatcherProvider.io) {
-                val cleanUrl = if (!url.startsWith("http://") && !url.startsWith("https://")) "http://$url" else url
-                val connection = URL(cleanUrl).openConnection() as HttpURLConnection
-                connection.connectTimeout = 5000
-                connection.readTimeout = 5000
-                connection.requestMethod = "GET"
-                val code = connection.responseCode
-                connection.disconnect()
-                code in 200..599
-            }
-        } catch (e: Exception) {
-            false
         }
     }
